開発試作/_archive/KindleAutoCap/requirements.txt
pyautogui
pillow
opencv-python 